Avoid Vec allocation for i2osp
In the vast majority of cases, we essentially want i2osp to act like
`to_be_bytes`. But, it does help to actually express things in a manner
consistent with the specs.  So, this makes i2osp a function with a const
generic parameter to control the size of the returned array.  It remains
as flexible as it was, but in the vast majority of cases where a usize
is turned into an 8-byte array, this compiles down to just a byte swap.